backend/pyproject.toml
[project]
name = "sitesearch-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "psycopg[binary]>=3.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "alembic>=1.14",
    "geoalchemy2>=0.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
